requirements.txt
matplotlib~=3.8.4
pyserial~=3.5